// src/app/api/(main)/lib/dataModel.ts
export interface HttpModel {
    url: string;
    method: Method;
    headers?: Record<string, string>;
    queryParams?: Record<string, string>;
    body?: any;
    description?: string;
}

export type Method =
    | 'get' | 'GET'
    | 'delete' | 'DELETE'
    | 'head' | 'HEAD'
    | 'options' | 'OPTIONS'
    | 'post' | 'POST'
    | 'put' | 'PUT'
    | 'patch' | 'PATCH'
    | 'purge' | 'PURGE'
    | 'link' | 'LINK'
    | 'unlink' | 'UNLINK';